fix some problem in user CRUD and add some permission in it

+++ Backend/api/views.py
@@ -4,6 +4,7 @@ from rest_framework.decorators import api_view
 #from .models import User
 from .serializers import UserSerializer
 from django.contrib.auth.models import User
+from django.contrib import auth
 import datetime
 # Create your views here.
 
@@ -24,19 +25,21 @@ def user_account(request):
             message = {'msg':'account already exist:','time':datetime.datetime.now()}
             return Response(message)
         except:
-            user_data = User.objects.create_user(data['account'],data['password'],data['email'])
+            user_data = User.objects.create_user(data['account'],data['email'],data['password'])
             serializer = UserSerializer(user_data,many=False)
             message = {'msg':'create success','detail':serializer.data}
             return Response(message)
 
     elif request.method == 'GET':
+        data = request.data
         try:
-            data = request.data
             user_data = User.objects.get(username = data['account'])
+            print(user_data)
             serializer = UserSerializer(user_data,many=False)
-            return Response(serializer.data['account'])
+            message = {'msg':'query success user: '+str(serializer.data['username'])+" is exist" ,'time':datetime.datetime.now()}
+            return Response(message)
         except:
-            message = {'msg':'query error:'+str(data) ,'time':datetime.datetime.now()}
+            message = {'msg':'query error or account '+data['account'] +' not exist' ,'time':datetime.datetime.now()}
             return Response(message)
     elif request.method == 'PUT':
         data = request.data
@@ -45,17 +48,22 @@ def user_account(request):
             user_data.set_password(data['password'])
             user_data.save()
             serializer = UserSerializer(user_data,many=False)
-            return Response(serializer.data)
+            message = {'msg':'modify success!','time':datetime.datetime.now()}
+            return Response(message)
         except:
             message = {'msg':'query error:'+str(data) ,'time':datetime.datetime.now()}
             return Response(message)
     elif request.method =='DELETE':
         data = request.data
         try:
-            user_data = User.objects.get(username = data['account'])
-            user_data.delete()
-            message = {'msg':'delete account success','time':datetime.datetime.now()}
-            return Response(message)
+            if request.user.is_superuser:
+                user_data = User.objects.get(username = data['account'])
+                user_data.delete()
+                message = {'msg':'delete account success','time':datetime.datetime.now()}
+                return Response(message)
+            else:
+                message = {'msg':'permission denied ','time':datetime.datetime.now()}
+                return Response(message)
         except:
             message = {'msg':'query error:'+str(data) ,'time':datetime.datetime.now()}
             return Response(message)
@@ -66,12 +74,18 @@ def user_account(request):
         
 @api_view(['GET'])
 def all_user(request):
-    print(request.META['REMOTE_ADDR'],request.user)
-    try:
+    print(request.META['REMOTE_ADDR'],'\n',request.user)
+    if request.user.is_superuser:
         check_data = User.objects.get(username = request.user)
         user_data = User.objects.all()
         serializer = UserSerializer(user_data,many=True)
         return Response(serializer.data)
-    except:
+    else:
         message = {"msg":"I can't tell you. You are "+str(request.user),'time':datetime.datetime.now()}
         return Response(message)
+'''
+@api_view(['POST'])
+def login(request):
+    if request.user.is_authenticated():
+        return 
+'''
